Use a single build file, with dev/production check. Use html-minifier for production
